requirements.txt
streamlit
yfinance
langchain
langgraph
langchain-google-genai
tavily-python
python-dotenv
